requirements.txt
flask==2.3.3
flask-cors==4.0.0
pytubefix
openai-whisper
numpy==1.26.3
ffmpeg-python==0.2.0
setuptools==68.2.2
--index-url https://download.pytorch.org/whl/cu121
torch==2.2.1+cu121 